Changed some critical parts of rply from type long to type int, so that it works on both 32-bit and 64-bit systems (before, it probably worked on 16-bit and 32-bit systems).

+++ tools/rply/rply.c
@@ -1185,16 +1185,16 @@ static int ply_type_check(void) {
     assert(sizeof(unsigned char) == 1);
     assert(sizeof(short) == 2);
     assert(sizeof(unsigned short) == 2);
-    assert(sizeof(long) == 4);
-    assert(sizeof(unsigned long) == 4);
+    assert(sizeof(int) == 4);
+    assert(sizeof(unsigned int) == 4);
     assert(sizeof(float) == 4);
     assert(sizeof(double) == 8);
     if (sizeof(char) != 1) return 0;
     if (sizeof(unsigned char) != 1) return 0;
     if (sizeof(short) != 2) return 0;
     if (sizeof(unsigned short) != 2) return 0;
-    if (sizeof(long) != 4) return 0;
-    if (sizeof(unsigned long) != 4) return 0;
+    if (sizeof(int) != 4) return 0;
+    if (sizeof(unsigned int) != 4) return 0;
     if (sizeof(float) != 4) return 0;
     if (sizeof(double) != 8) return 0;
     return 1;
@@ -1224,12 +1224,12 @@ static int oascii_uint16(p_ply ply, double value) {
 }
 
 static int oascii_int32(p_ply ply, double value) {
-    if (value > LONG_MAX || value < LONG_MIN) return 0;
+    if (value > INT_MAX || value < INT_MIN) return 0;
     return fprintf(ply->fp, "%d ", (int) value) > 0;
 }
 
 static int oascii_uint32(p_ply ply, double value) {
-    if (value > ULONG_MAX || value < 0) return 0;
+    if (value > UINT_MAX || value < 0) return 0;
     return fprintf(ply->fp, "%d ", (unsigned int) value) > 0;
 }
 
@@ -1268,14 +1268,14 @@ static int obinary_uint16(p_ply ply, double value) {
 }
 
 static int obinary_int32(p_ply ply, double value) {
-    long int32 = (long) value;
-    if (value > LONG_MAX || value < LONG_MIN) return 0;
+    int int32 = (int) value;
+    if (value > INT_MAX || value < INT_MIN) return 0;
     return ply->odriver->ochunk(ply, &int32, sizeof(int32));
 }
 
 static int obinary_uint32(p_ply ply, double value) {
-    unsigned long uint32 = (unsigned long) value;
-    if (value > ULONG_MAX || value < 0) return 0;
+    unsigned int uint32 = (unsigned int) value;
+    if (value > UINT_MAX || value < 0) return 0;
     return ply->odriver->ochunk(ply, &uint32, sizeof(uint32));
 }
 
@@ -1328,7 +1328,7 @@ static int iascii_int32(p_ply ply, double *value) {
     char *end;
     if (!ply_read_word(ply)) return 0;
     *value = strtol(BWORD(ply), &end, 10);
-    if (*end || *value > LONG_MAX || *value < LONG_MIN) return 0;
+    if (*end || *value > INT_MAX || *value < INT_MIN) return 0;
     return 1;
 }
 
@@ -1385,14 +1385,14 @@ static int ibinary_uint16(p_ply ply, double *value) {
 }
 
 static int ibinary_int32(p_ply ply, double *value) {
-    long int32;
+    int int32;
     if (!ply->idriver->ichunk(ply, &int32, sizeof(int32))) return 0;
     *value = int32;
     return 1;
 }
 
 static int ibinary_uint32(p_ply ply, double *value) {
-    unsigned long uint32;
+    unsigned int uint32;
     if (!ply->idriver->ichunk(ply, &uint32, sizeof(uint32))) return 0;
     *value = uint32;
     return 1;
